+++ WizardOfLegend/WizardOfLegend/Ui/PlayUiMgr.h
@@ -17,8 +17,8 @@ protected:
 	float maxOverdriveBarSize = 48.f * 4.f;
 	float overdriveBarSize = 0.f;
 
-	float maxHpBarSize = 62.f * 4.f;
-	float hpBarSize = 62.f * 4.f;
+	float maxHpBarSize = 62 * 4;
+	float hpBarSize = 62 * 4;
 	float hpBarHurtSize = 62.f * 4.f;
 
 	bool testOverdrive = false;
@@ -26,8 +26,8 @@ protected:
 
 	// 플레이어한테 받아와야 함
 	TextObj* hpText;
-	float playerCurHp = 525.f;
-	float playerMaxHp = 525.f;
+	int playerCurHp = 525;
+	int playerMaxHp = 525;
 
 public:
 	PlayUiMgr();
